requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.23.1
tensorflow==2.12.0
numpy==1.24.3
pillow==9.5.0
python-multipart==0.0.6
